Replace integer casts with double casts, to fix crash with invalid cast. This issue arises because gson tries to guess a number type and uses double by default for most compatability. Unfortunatly we cannot update gson to support setting there default values

+++ src/main/java/info/u_team/gltf_parser/BinaryGLTFParser.java
@@ -70,7 +70,7 @@ public class BinaryGLTFParser extends GLTFParser {
 	
 	@Override
 	public ByteBuffer getData(BufferView bufferView) {
-		final int bufferIndex = (Integer) bufferView.getBuffer();
+		final int bufferIndex = ((Double) bufferView.getBuffer()).intValue();
 		if (bufferIndex != 0)
 			throw new UnsupportedOperationException("Index other then 0 not allowed in binary mode!");
 		final ByteBuffer data = ByteBuffer.wrap(gltfToBinaryData);
